Update the test folder structure to meet Single Responsibility Principle
