+++ lib/src/exceptions/conditionals/conditional_exceptions.dart
@@ -1,0 +1,4 @@
+/// Copyright (C) 2022 Intel Corporation
+/// SPDX-License-Identifier: BSD-3-Clause
+
+export 'signal_redriven_exception.dart';
