+++ lib/src/exceptions/exceptions.dart
@@ -1,6 +1,6 @@
 /// Copyright (C) 2022 Intel Corporation
 /// SPDX-License-Identifier: BSD-3-Clause
 
-export 'conditional_exceptions.dart';
-export 'name_exceptions.dart';
-export 'sim_compare/sim_compare_exceptions.dart';
+export './conditionals/conditional_exceptions.dart';
+export './name/name_exceptions.dart';
+export './sim_compare/sim_compare_exceptions.dart';
