+++ lib/src/exceptions/name/invalid_reserved_name_exception.dart
@@ -1,31 +1,13 @@
 /// Copyright (C) 2022 Intel Corporation
 /// SPDX-License-Identifier: BSD-3-Clause
 ///
-/// name_exception.dart
-/// Name Exception that have custom type to be thrown,
+/// invalid_reserved_name_exception.dart
+/// An exception that is thrown when reserved name is invalid
 ///
 /// 2022 October 25
 /// Author: Yao Jing Quek <yao.jing.quek@intel.com>
 ///
 import 'package:rohd/rohd.dart';
-
-/// This Exception show that reserved name eg. definitionName is NULL but
-/// reserve flag eg. reserveDefinitionName is set to True.
-///
-/// Please check on the class [Module] for the constructor argument.
-class NullReservedNameException implements Exception {
-  late final String _message;
-
-  /// constructor for NullReservedNameException,
-  /// pass custom message to the constructor
-  NullReservedNameException(
-      [String message = 'Reserved Name cannot be null '
-          'if reserved name set to true'])
-      : _message = message;
-
-  @override
-  String toString() => _message;
-}
 
 /// This Exception show that reserved name eg. definitionName naming convention
 /// is invalid but reserve flag eg. reserveDefinitionName is set to True.
@@ -36,7 +18,7 @@ class InvalidReservedNameException implements Exception {
   late final String _message;
 
   /// constructor for InvalidReservedNameException,
-  /// pass custom message to the constructor
+  /// pass custom String [message] to the constructor to override
   InvalidReservedNameException(
       [String message = 'Reserved Name need to follow proper naming '
           'convention if reserved'
