+++ lib/src/exceptions/name/name_exceptions.dart
@@ -1,0 +1,5 @@
+/// Copyright (C) 2022 Intel Corporation
+/// SPDX-License-Identifier: BSD-3-Clause
+
+export 'invalid_reserved_name_exception.dart';
+export 'null_reserved_name_exception.dart';
