+++ lib/src/modules/conditional.dart
@@ -12,7 +12,7 @@ import 'dart:async';
 
 import 'package:meta/meta.dart';
 import 'package:rohd/rohd.dart';
-import 'package:rohd/src/exceptions/conditional_exceptions.dart';
+import 'package:rohd/src/exceptions/conditionals/conditional_exceptions.dart';
 import 'package:rohd/src/types/duplicate_detection_set.dart';
 import 'package:rohd/src/utilities/sanitizer.dart';
 import 'package:rohd/src/utilities/uniquifier.dart';
